+++ src/components/layout/Banner/Banner.module.css
@@ -28,7 +28,7 @@
 
 .title {
     font-weight: 500;
-    font-size: 4.7rem;
+    font-size: 5.3rem;
     z-index: 3;
     color: #fff;
 }
