+++ src/components/layout/Banner/Banner.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import { Btn } from '../UI/Btn';
+import { Btn } from '../../UI/Btn';
 
 import styles from './Banner.module.css';
-import backgroundImg from '../../assets/car-background.png';
+import backgroundImg from '../../../assets/car-background.png';
 
 export const Banner = () => {
     return <>
         <div className={styles.bannerBox} style={{backgroundImage: `url(${backgroundImg})`}}>
             <div className={styles.dimness}/>
             <div className={styles.banner}>
-                <h1 className={styles.title}>Znajdziesz tu każdy<br/>samochód którego szukasz.</h1>
+                <h1 className={styles.title}>Samochody, które<br/>roztopią twoje uczucia.</h1>
                 <Btn to="/add" className={styles.button} btnStyle text="Dodaj ogłoszenie"/>
             </div>
         </div>
