+++ src/components/layout/NotFound/NotFound.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { Btn } from '../UI/Btn';
+import { Btn } from '../../UI/Btn';
 
 import styles from './NotFound.module.css';
 
 export const NotFound = () => (
     <div className={styles.notFound}>
-        <h1>Ups! Zabrakło nam paliwa...</h1>
+        <h1>Zwolnij kierowco! Zabrakło nam paliwa...</h1>
         <p>Przepraszamy, ale nie znaleziono strony o podanym adresie. <Btn to="/" text="Wróć na stronę główną."/></p>
     </div>
 );
